backend/requirements.txt
SQLAlchemy>=2.0.32
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
stripe>=9.5.0
pandas>=2.2.2
plotly>=5.24.1
streamlit>=1.38.0
tenacity>=9.0.0
requests>=2.32.3
pydantic>=2.8.2
alembic>=1.13.2
apscheduler>=3.10.4
Flask>=3.0.0

# Google APIs
google-analytics-data>=0.18.8
google-auth>=2.23.0
google-api-python-client>=2.139.0
google-ads>=24.1.0

